Changes:
Plugin.download verändert. Speedmessung läuft jetzt über speedmeter. Der laufende Durchschnitt wird dabei verwendet

HTTPReconnect.java
Kann jetzt über eine checkbox kurzzeitig deaktiviert werden

Controlling: Mehere Events hinzugefügt

Dropper:
setText infos weggenommen
TabDownloadslinks:
links können verschoben werden. 
Rechtsklick menü

PLugin.post/getRequest:
gzip unterstützung(nicht getestet)

Rapidshare.com:
Hardeware defekts eine eingebaut (nicht getestet)


git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@184 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/skins/simple/Speedometer.java
@@ -37,12 +37,12 @@ public class Speedometer extends JPanel{
         this.downloadLink = downloadLink;
         lblDownloadName.setText(downloadLink.getName());
         if (downloadLink.getDownloadMax()>0)
-            lblDownloadSize.setText(Integer.toString(downloadLink.getDownloadMax()));
+            lblDownloadSize.setText(Long.toString(downloadLink.getDownloadMax()));
     }
 
     private class SpeedometerRefresher extends Thread{
         private final int REFRESH_TIME = 1000;
-        private int downloadedBytes=0;
+        private long downloadedBytes=0;
         
         public SpeedometerRefresher(){
             super("JD-SpeedometerRefresher");
@@ -52,7 +52,7 @@ public class Speedometer extends JPanel{
             long differenceBytes;
             long differenceTime;
             long start,end;
-            
+           
             while(downloadLink!=null && downloadLink.isInProgress()){
                 start = System.currentTimeMillis();
                 downloadedBytes = downloadLink.getDownloadCurrent();
